--- app/src/main/java/yaroslav/ovdiienko/idivision/fragmentstest/imageprocessing/util/Mode.kt
@@ -1,7 +1,0 @@
-package yaroslav.ovdiienko.idivision.fragmentstest.imageprocessing.util
-
-
-enum class Mode {
-    HUE,
-    SATURATION
-}
