+++ app/src/main/java/yaroslav/ovdiienko/idivision/renderscripttest/imageprocessing/util/Mode.kt
@@ -1,0 +1,7 @@
+package yaroslav.ovdiienko.idivision.renderscripttest.imageprocessing.util
+
+
+enum class Mode {
+    HUE,
+    SATURATION
+}
