+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#1F1E1E"
-    tools:context=".MainActivity">
+    tools:context="yaroslav.ovdiienko.idivision.renderscripttest.MainActivity">
 
     <TextView
         android:id="@+id/textView"
