+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">FragmentsTest</string>
+    <string name="app_name">RenderScript</string>
 </resources>
